Alerts v2 (#42)
This PR completely changes the way we do alerts. (Note, however, that 5,500 lines of the diff is from a generated rules file!)

Instead of generating a custom alerting and recording rules file from the code, we now use a single file for all autometrics libraries.

The way this works is that the rules file defines a bunch of recording and alerting rules that are, by default, dormaint. The rules are enabled when specific labels are attached to some of the metrics produced by autometrics.

This massively simplifies supporting alerting across the various autometrics implementations because we only need to have them attach certain labels to the metrics, instead of needing to reimplement the alert generation logic in every library.

Thanks to @hatchan and @keturiosakys for their help rethinking how alerting should work and on the library API.


* Kill prometheus server when example stops

* Add autometrics-cli to generate Sloth SLO file

* Identify SLOs by name instead of number

* Refactor labels to use structs with named fields

* Alerts v2

* Remove alerts example

* Update readme

* Mark alerting rules as generated

* Fix glob

* Remove unused alerts keyword

* Move recording rules to root

* Remove empty file

* Don't panic on drop

* Fix gitattributes file

* Minor readme tweaks

* Add SLO example to readme